Avoid double copy in string
So we have a vector whihc we copy the TAG from the file into. Then we are reading that byte sequence as if it were UTF-8 and getting a view to that. Then we need to transfer that to a different scope so we are cloning it. This is not necessary since we no longer need that vector we should just move it.